Cast time_t values to int64_t
Needed to support platforms which has a time_t in 32 and 64 bit.
When working with time internally, time_t is fine, but using time_t
can be difficult when it changes size depending on platform.

+++ src/BPLSettings.cpp
@@ -429,7 +429,7 @@ bool BPLSettings::dejsonBeerProfile(String json)
 	const char *uintStr=doc["u"];
 	tempSchedule->unit=  *uintStr;
 
-	DBG_PRINTF("Load finished, st:%lld, unit:%c, _numberOfSteps:%d\n",tempSchedule->startDay,
+	DBG_PRINTF("Load finished, st:%lld, unit:%c, _numberOfSteps:%d\n",static_cast<std::int64_t>(tempSchedule->startDay),
 	tempSchedule->unit,tempSchedule->numberOfSteps);
 
 	return true;
