+++ src/BrewKeeper.cpp
@@ -207,7 +207,10 @@ float BrewProfile::tempByTimeGravity(time_t time,Gravity gravity)
 {
 	if(time < _schedule->startDay) return INVALID_CONTROL_TEMP;
 
-	DBG_PRINTF("currentStep:%d, timeEnterCurrentSTep:%lld, time:%lld\n",_status->currentStep,_status->timeEnterCurrentStep,time);
+	DBG_PRINTF("currentStep:%d, timeEnterCurrentSTep:%lld, time:%lld\n",
+	           _status->currentStep,
+	           static_cast<std::int64_t>(_status->timeEnterCurrentStep),
+	           static_cast<std::int64_t>(time));
 
 	if(	_status->startingDate ==0 
 	   || _status->startingDate != _schedule->startDay
@@ -216,8 +219,11 @@ float BrewProfile::tempByTimeGravity(time_t time,Gravity gravity)
 	}
 	if(_status->currentStep >= _schedule->numberOfSteps) return INVALID_CONTROL_TEMP;
 
-	DBG_PRINTF("tempByTimeGravity:now:%lld, step:%d, type=%c, last elapsed:%lld\n",time,
-		_status->currentStep,_schedule->steps[_status->currentStep].condition,time - _status->timeEnterCurrentStep);
+	DBG_PRINTF("tempByTimeGravity:now:%lld, step:%d, type=%c, last elapsed:%lld\n",
+	           static_cast<std::int64_t>(time),
+	           _status->currentStep,
+	           _schedule->steps[_status->currentStep].condition,
+	           static_cast<std::int64_t>(time - _status->timeEnterCurrentStep));
 
     if(checkCondition(time,gravity)){
     		// advance to next stage
